changed power_consumed variable and associated functions to type double, to match the power_contained variable in the Battery class

+++ Arm.cpp
@@ -1,11 +1,11 @@
 #include "Arm.h"
 
-Arm::Arm(string n, int pn, double w, double c, string d, int pc) : Robot_Part(n, pn, w, c, d, Component_type::Arm)
+Arm::Arm(string n, int pn, double w, double c, string d, double pc) : Robot_Part(n, pn, w, c, d, Component_type::Arm)
 {
 	power_consumed = pc;
 }
 
-int Arm::get_power_consumed()
+double Arm::get_power_consumed()
 {
 	return power_consumed;
 }
